requirements.txt
#Data Squad & Core Libraries
pandas
xarray
netcdf4
sqlalchemy
psycopg2-binary

#AI Squad Libraries
langchain
langchain-community
langchain-google-genai
sentence-transformers

#Backend API Libraries
fastapi
uvicorn[standard]
websockets
python-multipart
python-dotenv

#Frontend Squad Libraries
plotly
